Make Coord only manage units & not pixels
Having two coordinates systems is a pain to handle, thus this is
reduced to 1. Only units will be managed, and converted to pixels when
showing on screen. To manage sub-units coord, they are now done in
doubles instead of ints.

+++ movement.h
@@ -40,7 +40,7 @@ struct Movement
     Coord path[MAX_PATH_NODES]; // array of nodes
     unsigned int current_node;
     MovementType movement_type;
-    unsigned int velocity; // in pixels
+    double velocity; // in units
 };
 
 // prototypes
